change a test in vector and matrics(add a float)

+++ project/matrics.py
@@ -3,22 +3,22 @@ from typing import List, Optional
 
 def sum_matr(a: List[List[float]], b: List[List[float]]) -> Optional[List[List[float]]]:
 
-    return [[a[i][j] + b[i][j] for j in range(len(a[0]))] for i in range(len(a))]
+    return [[float(a[i][j] + b[i][j]) for j in range(len(a[0]))] for i in range(len(a))]
 
 
 def pro_matr(a: List[List[float]], b: List[List[float]]) -> Optional[List[List[float]]]:
     if len(a[0]) != len(b):
         return None
 
-    result = [[0 for _ in range(len(b[0]))] for _ in range(len(a))]
+    result = [[0.0 for _ in range(len(b[0]))] for _ in range(len(a))]
     for i in range(len(a)):
         for j in range(len(b[0])):
             for k in range(len(b)):
-                result[i][j] += a[i][k] * b[k][j]
+                result[i][j] += float(a[i][k] * b[k][j])
 
     return result
 
 
 def tra_matr(a: List[List[float]]) -> List[List[float]]:
 
-    return [[a[j][i] for j in range(len(a))] for i in range(len(a[0]))]
+    return [[float(a[j][i]) for j in range(len(a))] for i in range(len(a[0]))]
